- change vpmask API into new vpset API where bit=1 means corresponding VP is selected and bit=0 means corresponding VP is not selected, see marcel_topology.h
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@18154 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ marcel/include/asm-alpha/linux_bitops.h
@@ -391,11 +391,11 @@ static __tbx_inline__ unsigned long __ma_ffs(unsigned long word)
  */
 
 #section marcel_functions
-static __tbx_inline__ int ma_ffs(int word);
+static __tbx_inline__ unsigned long ma_ffs(unsigned long word);
 #section marcel_inline
-static __tbx_inline__ int ma_ffs(int word)
+static __tbx_inline__ unsigned long ma_ffs(unsigned long word)
 {
-	int result = __ma_ffs(word) + 1;
+	unsigned long result = __ma_ffs(word) + 1;
 	return word ? result : 0;
 }
